* Fix the issue of transfer context being inject into wrong address   as the page offset was some-how not considered
* Fix the refactoring and various compile time error

* Adjust the lunadbg to work with latest threading refactoring.

* Also fix the issue that lunadbg's llist iterator had made false
  assumption on the offset of embeded llist_header.

* Rename spawn_thread -> create_thread. And introduce spawn_kthread
  to spawn a kernel thread within kernel process.

* Fix the issue in vmm_lookupat that ignore the present bit when
  doing pte probing

* Leaves some holes for later investigations

+++ lunaix-os/kernel/mm/mmap.c
@@ -141,7 +141,7 @@ found:
         attr &= ~PG_ALLOW_USER;
     }
 
-    for (int i = 0; i < param->mlen; i += PG_SIZE) {
+    for (size_t i = 0; i < param->mlen; i += PG_SIZE) {
         vmm_set_mapping(param->vms_mnt, found_loc + i, 0, attr, 0);
     }
